Prevent JDWP thread death when adbd dies.
The handling for "remote side disconnected" was getting folded in with
more severe failures.  This separates the two, returning to the connection
retry loop when adbd disappears.

One side-effect is that the JDWP thread will now wake up every couple of
seconds when adbd isn't running.  Only affects "debuggable" apps.

This also reduces the priority on a couple of log messages.

For internal bug 1844156.