Inline some java.lang.Math functions.
For a first pass, I inlined the various flavors of abs(), min()/max() on
integers, sqrt(), cos(), and sin().  These were selected based on a
static analysis of a few of our jar files.

A test of repeated sin/cos/sqrt calls on a G1-class device showed an
improvement of 28%.  This would improve more on devices with VFP
support if the VM is compiled with -mfpu=vfp.

Also: clarified a warning and removed some "#if 0" stuff.