Collation.getSortKey returns keys generated by a different algorithm
This test tested implementation details. This change makes the test only check specified behavior.

Note: The resulting key is incompatible with a key returned by the RI. that's the downside of using icu.
But there's also an upside: ICU compresses the sorting key that gets returned. This increases performance
when comparing the keys.

BUG=1635900