Avoid JDWP hang on nested thread suspension.
With jdb, you can suspend a thread multiple times.  If you try to
execute a method -- which requires resuming a thread that has been
stopped at a breakpoint -- the VM currently only does a single "resume",
which means the thread is still suspended, and the JDWP thread hangs
waiting for it.

This adds a check to prevent the hang.

For bug 2183735.