Iterate in JNI thread exit check.
The Dalvik VM includes a helpful check that kicks and screams if a thread
exits without detaching from the VM first.  It manages this with a pthread
TLS destructor.  This is handy, but it will interfere with anyone who
wants to use a TLS destructor to initiate the thread detach.

With this change, we now iterate a couple of times to allow other
destructors a chance to detach the thread before we get snippy.

For 2319072.