Add instructions for volatile wide fields.
This adds four new instructions for accessing volatile wide fields (long
and double).  The JLS requires that such accesses are atomic, but the
VM doesn't otherwise make guarantees about the atomicity of reads and
writes on 64-bit fields.

There are no behavioral changes.  This just adds definitions for the new
instructions and a couple of tests.  The current implementation is just
the non-volatile form of the instructions or a C stub, but since we're
not generating them it doesn't really matter yet.

Also:
 - bumped Dalvik version to 1.3.0
 - added a note to the x86-atom TODO list

For bug 1633591.