Increase sleep delays on debug thread kill.
It appears the delays between the signals were insufficient for a busy
system.  This increases the timeouts to larger values, which may or may
not be sufficient.  What's really needed here is a way to monitor
debuggerd's activity, but that seems like a lot of work for what this
wants to accomplish.

Also, set gDvm.nativeDebuggerActive to true to prevent the heapworker
watchdog timeout from firing.  (Either somebody is going to attach a
debugger soon, or the process is going to die, so it's reasonable to
do here.)

Change-Id: Ic92590499871860ef766147eeaa6efc7afb4acdc