Fix all our java.text.Bidi failures.
Two are invalid tests where we behave the same as the RI. The third reflects
a difference between ICU4C and the RI, so I've added code to recognize that
case and paper over the crack. I've also tidied up a bit more.

With any luck, I'll never see these files again...

Change-Id: I278502fbdea039e9a2318df830dc0b55851e96c0