Tidying up the error reporting.
We were spamming our logs with a bunch of redundant messages,
particularly surrounding timeouts. This should reduce the log
output to only the useful stuff when a test takes too long
to complete.

Also not blowing away the full XML report when a few tests
get lost along the way.