Add native method arguments to GC scan.
We want to identify arguments passed to native methods (internal native
or JNI) so that we can tell the GC to leave those blocks alone.  The
alternative is to add explicit "add to tracked ref list" calls to a
bunch of internal natives and to the JNI call bridge.

We walk through the shorty signature to make this a precise scan.

Since something is now looking at the method arguments, it was necessary
to actually set a value in the argument area on the stack for the "fake"
native method entry points used for the main thread and for externally-
created threads that are later attached to the VM.

(The implementation is currently #ifdefed-out, since there's nothing
for it to do yet.)

Change-Id: I2143ea541bc2fc50e81d36ad1c251b041553eaef