Properly relocate objects with exposed identity hash codes.
When the transporter encounters an identity hashed object it will
allocate an additional word in to-space that it fills with the
identity hash code.  The identity has code is the from-space address
for uncopied objects and a word immediately after the instance for
copied objects.  The transporter will update the hash code state when
it relocates the identity hash code.

All size computations are now performed by objectSize which now
interprets the hash code state.  Similarly, all transportation logic
has been moved to transportObject which knows how to allocate and fill
the word holding the identity hash code.

Change-Id: I4920e18299850168b41158de748c8331be541ca1