Fix the identity hash code getter.
The address of the hash code for array objects was incorrectly rounded
up to the next double word instead of the next word as the comment
suggested.  Also, assert that we are not in receipt of a class object.
Class objects should not be in the hashed-and-moved state as they are
not yet relocatable.

Change-Id: I0d2ee147f3c144d7308e80aea32fc3dd1719402e