Jit: Fix for  2717275 First JIT bug reported by external developers
The JIT was incorrectly keeping a short value in a floating point
register rather than copying it to a core register before storing.
There was an assert to catch this case, but asserts don't fire in
production builds.

The fix is safe and simple - just exclude this case from the "optimization".

Change-Id: I38f1865da11ef3f78213f84967d4737e066b9c68