Jit: Fix for  2717275 First JIT bug reported by external developers
The JIT was incorrectly keeping a short value in a floating point
register rather than copying it to a core register before storing.
There was an assert to catch this case, but asserts don't fire in
production builds.

The fix is safe and simple - just exclude this case from the "optimization".

Change-Id: I33767c8a202b6fa36a19d918ac5b914a5e4e4de3