Add a callback driven visitor routine and make verify use it.
This is the first step along the way to ending the proliferation of
one-off object walking routines.  This code should supersede similar
code in the mark sweep collector, the copying collector, the hprof
tool, etc..  Once all of its uses have been debugged the routines in
the .c file should be hoisted into the .h file and marked as inline.

Soon, the signature of the visitor will probably have to be converted
to a functor with members specialized to reference type and, in
addition, given an extra parameter for a context structure.

Change-Id: Ib00850f22374c5aca51fa3708304774da90d8bc4