Eliminate pointer tagging on the reference operation queue.
Tags are unconditionally added before references are added to the
reference queue.  This tag is checked when a reference is removed from
the queue and subsequently cleared.  There are no other observers of
this tag and so it is equivalent to not tag the reference in the first
place.  The original intent may have been to tag the references in
different ways.  Removing this code simplifies reference tracing.

Change-Id: Idb18929e601acdd049b4a92ebf588f3f5febb75c